requirements.txt
torch
transformers
datasets
pyyaml
matplotlib
huggingface_hub
tqdm
openai 
